+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.0] - 2021-10-15
+
 ### Removed
 
 - Support for the old sled database.
@@ -274,7 +276,8 @@ It is possible to migrate critical data from the old db to the sqlite but there 
 - Fixed an issue where Alice would not verify if Bob's Bitcoin lock transaction is semantically correct, i.e. pays the agreed upon amount to an output owned by both of them.
   Fixing this required a **breaking change** on the network layer and hence old versions are not compatible with this version.
 
-[Unreleased]: https://github.com/comit-network/xmr-btc-swap/compare/0.9.0...HEAD
+[Unreleased]: https://github.com/comit-network/xmr-btc-swap/compare/0.10.0...HEAD
+[0.10.0]: https://github.com/comit-network/xmr-btc-swap/compare/0.9.0...0.10.0
 [0.9.0]: https://github.com/comit-network/xmr-btc-swap/compare/0.8.3...0.9.0
 [0.8.3]: https://github.com/comit-network/xmr-btc-swap/compare/0.8.2...0.8.3
 [0.8.2]: https://github.com/comit-network/xmr-btc-swap/compare/0.8.1...0.8.2